Rename Command & Parameter to include Node.
This is how they were named earlier but I had changed it as at that
time, I planned that people would interact more directly with those
classes.

That plan has changed now though and the higher level interaction
will be with a new Command class that is part of the CommandTable
infrastructure.